Find corresponding `mentions-popup` instead first
When you have multiple inputs on a page (for example the discussion view page, where you can have a `add` and `edit` action on one page), the mentions-popup always used the first one found to show the results. This causes problems, where sometimes it isn't even visible due the textarea is being hidden.